+++ Editor/EditorCore/EditorSceneViewPicker.cs
@@ -15,7 +15,9 @@ namespace UnityEditor.ProBuilder
         public const float offPointerMultiplier = 1.2f;
 
         public CullingMode cullMode;
+#if !UNITY_2023_2_OR_NEWER
         public SelectionModifierBehavior selectionModifierBehavior;
+#endif
         public RectSelectMode rectSelectMode;
     }
 
@@ -270,7 +272,9 @@ namespace UnityEditor.ProBuilder
             };
 
             UndoUtility.RecordSelection("Drag Select");
-            bool isAppendModifier = EditorHandleUtility.IsAppendModifier(Event.current.modifiers);
+            bool isSelectionAddModifier = EditorHandleUtility.IsSelectionAddModifier(Event.current.modifiers);
+            bool isSelectionRemoveModifier = EditorHandleUtility.IsSelectionAppendOrRemoveIfPresentModifier(Event.current.modifiers);
+            bool isAppendModifier = isSelectionAddModifier || isSelectionRemoveModifier;
 
             if (!isAppendModifier)
                 MeshSelection.ClearElementSelection();
@@ -299,12 +303,19 @@ namespace UnityEditor.ProBuilder
                         {
                             common = mesh.GetSharedVertexHandles(mesh.selectedIndexesInternal);
 
+#if UNITY_2023_2_OR_NEWER
+                            if(isSelectionAddModifier)
+                                common.UnionWith(kvp.Value);
+                            else if(isSelectionRemoveModifier)
+                                common.RemoveWhere(x => kvp.Value.Contains(x));
+#else
                             if (scenePickerPreferences.selectionModifierBehavior  == SelectionModifierBehavior.Add)
                                 common.UnionWith(kvp.Value);
                             else if (scenePickerPreferences.selectionModifierBehavior  == SelectionModifierBehavior.Subtract)
                                 common.RemoveWhere(x => kvp.Value.Contains(x));
                             else if (scenePickerPreferences.selectionModifierBehavior  == SelectionModifierBehavior.Difference)
                                 common.SymmetricExceptWith(kvp.Value);
+#endif
                         }
                         else
                         {
@@ -335,13 +346,19 @@ namespace UnityEditor.ProBuilder
                         if (isAppendModifier)
                         {
                             current = new HashSet<Face>(kvp.Key.selectedFacesInternal);
-
+#if UNITY_2023_2_OR_NEWER
+                            if(isSelectionAddModifier)
+                                current.UnionWith(kvp.Value);
+                            else if(isSelectionRemoveModifier)
+                                current.RemoveWhere(x => kvp.Value.Contains(x));
+#else
                             if (scenePickerPreferences.selectionModifierBehavior == SelectionModifierBehavior.Add)
                                 current.UnionWith(kvp.Value);
                             else if (scenePickerPreferences.selectionModifierBehavior == SelectionModifierBehavior.Subtract)
                                 current.RemoveWhere(x => kvp.Value.Contains(x));
                             else if (scenePickerPreferences.selectionModifierBehavior == SelectionModifierBehavior.Difference)
                                 current.SymmetricExceptWith(kvp.Value);
+#endif
                         }
                         else
                         {
@@ -375,13 +392,19 @@ namespace UnityEditor.ProBuilder
                         if (isAppendModifier)
                         {
                             current = EdgeLookup.GetEdgeLookupHashSet(mesh.selectedEdges, common);
-
+#if UNITY_2023_2_OR_NEWER
+                            if(isSelectionAddModifier)
+                                current.UnionWith(selectedEdges);
+                            else if(isSelectionRemoveModifier)
+                                current.RemoveWhere(x => selectedEdges.Contains(x));
+#else
                             if (scenePickerPreferences.selectionModifierBehavior == SelectionModifierBehavior.Add)
                                 current.UnionWith(selectedEdges);
                             else if (scenePickerPreferences.selectionModifierBehavior == SelectionModifierBehavior.Subtract)
                                 current.RemoveWhere(x => selectedEdges.Contains(x));
                             else if (scenePickerPreferences.selectionModifierBehavior == SelectionModifierBehavior.Difference)
                                 current.SymmetricExceptWith(selectedEdges);
+#endif
                         }
                         else
                         {