+++ Editor/EditorCore/ProBuilderEditor.cs
@@ -80,8 +80,10 @@ namespace UnityEditor.ProBuilder
         static Pref<RectSelectMode> m_DragSelectRectMode =
             new Pref<RectSelectMode>("editor.dragSelectRectMode", RectSelectMode.Partial);
 
+#if !UNITY_2023_2_OR_NEWER
         static Pref<SelectionModifierBehavior> m_SelectModifierBehavior =
             new Pref<SelectionModifierBehavior>("editor.rectSelectModifier", SelectionModifierBehavior.Difference);
+#endif
 
         internal static RectSelectMode rectSelectMode
         {
@@ -99,6 +101,7 @@ namespace UnityEditor.ProBuilder
             }
         }
 
+#if !UNITY_2023_2_OR_NEWER
         internal static SelectionModifierBehavior selectionModifierBehavior
         {
             get { return m_SelectModifierBehavior.value; }
@@ -114,6 +117,7 @@ namespace UnityEditor.ProBuilder
                     s_Instance.m_ScenePickerPreferences.selectionModifierBehavior = value;
             }
         }
+#endif
 
         internal static bool backfaceSelectionEnabled
         {
@@ -361,7 +365,9 @@ namespace UnityEditor.ProBuilder
             m_ScenePickerPreferences = new ScenePickerPreferences()
             {
                 cullMode = m_BackfaceSelectEnabled ? CullingMode.None : CullingMode.Back,
+#if !UNITY_2023_2_OR_NEWER
                 selectionModifierBehavior = m_SelectModifierBehavior,
+#endif
                 rectSelectMode = m_DragSelectRectMode
             };
         }
@@ -472,7 +478,12 @@ namespace UnityEditor.ProBuilder
             if (m_CurrentEvent.type == EventType.KeyDown)
             {
                 // Escape isn't assignable as a shortcut
+#if UNITY_2023_2_OR_NEWER
+                //Do not escape the select mode if an active tool override is active
+                if (m_CurrentEvent.keyCode == KeyCode.Escape && selectMode != SelectMode.Object && EditorToolManager.activeOverride == null)
+#else
                 if (m_CurrentEvent.keyCode == KeyCode.Escape && selectMode != SelectMode.Object)
+#endif
                 {
                     selectMode = SelectMode.Object;
 
@@ -567,7 +578,7 @@ namespace UnityEditor.ProBuilder
 
         internal void HandleMouseEvent(SceneView sceneView, int controlID)
         {
-            if(m_CurrentEvent.type == EventType.MouseDown && HandleUtility.nearestControl == controlID)
+            if(m_CurrentEvent.type == EventType.MouseDown && m_CurrentEvent.keyCode == KeyCode.Mouse0 && HandleUtility.nearestControl == controlID)
             {
                 // double clicking object
                 if(m_CurrentEvent.clickCount > 1)
@@ -891,7 +902,9 @@ namespace UnityEditor.ProBuilder
                         break;
                 }
 
-                selectMode = UI.EditorGUIUtility.DoElementModeToolbar(m_ElementModeToolbarRect, selectMode);
+                var mode = UI.EditorGUIUtility.DoElementModeToolbar(m_ElementModeToolbarRect, selectMode);
+                if (selectMode != mode)
+                    selectMode = mode;
 
                 if (s_ShowSceneInfo)
                 {