+++ Editor/MenuActions/Editors/OpenVertexColorEditor.cs
@@ -6,7 +6,7 @@ namespace UnityEditor.ProBuilder.Actions
     sealed class OpenVertexColorEditor : MenuAction
     {
         public override ToolbarGroup group { get { return ToolbarGroup.Tool; } }
-        public override Texture2D icon { get { return IconUtility.GetIcon("Toolbar/Panel_VertColors", IconSkin.Pro); } }
+        public override Texture2D icon { get { return IconUtility.GetIcon("Toolbar/Panel_VertColors"); } }
         public override TooltipContent tooltip { get { return s_Tooltip; } }
         public override string menuTitle { get { return "Vertex Colors"; } }