+++ Editor/MenuActions/Export/Export.cs
@@ -5,13 +5,17 @@ using System.Collections.Generic;
 using UnityEngine.ProBuilder.Stl;
 using UnityEngine.ProBuilder;
 using Object = UnityEngine.Object;
+#if UNITY_2023_2_OR_NEWER
+using UnityEditor.Actions;
+using UnityEngine.UIElements;
+#endif
 
 namespace UnityEditor.ProBuilder.Actions
 {
     sealed class Export : MenuAction
     {
         public override ToolbarGroup group { get { return ToolbarGroup.Object; } }
-        public override Texture2D icon { get { return IconUtility.GetIcon("Toolbar/Object_Export", IconSkin.Pro); } }
+        public override Texture2D icon { get { return IconUtility.GetIcon("Toolbar/Object_Export"); } }
         public override TooltipContent tooltip { get { return m_Tooltip; } }
         protected override bool hasFileMenuEntry { get { return false; } }