+++ Editor/MenuActions/Geometry/DeleteFaces.cs
@@ -12,10 +12,7 @@ namespace UnityEditor.ProBuilder.Actions
             get { return ToolbarGroup.Geometry; }
         }
 
-        public override Texture2D icon
-        {
-            get { return IconUtility.GetIcon("Toolbar/Face_Delete", IconSkin.Pro); }
-        }
+        public override Texture2D icon { get { return IconUtility.GetIcon("Toolbar/Face_Delete"); } }
 
         public override TooltipContent tooltip
         {