+++ Editor/MenuActions/Object/FlipObjectNormals.cs
@@ -1,8 +1,5 @@
 using UnityEngine.ProBuilder;
-using UnityEditor.ProBuilder;
 using UnityEngine;
-using UnityEditor;
-using UnityEditor.ProBuilder.UI;
 
 namespace UnityEditor.ProBuilder.Actions
 {
@@ -13,10 +10,7 @@ namespace UnityEditor.ProBuilder.Actions
             get { return ToolbarGroup.Object; }
         }
 
-        public override Texture2D icon
-        {
-            get { return IconUtility.GetIcon("Toolbar/Object_FlipNormals", IconSkin.Pro); }
-        }
+        public override Texture2D icon { get { return IconUtility.GetIcon("Toolbar/Object_FlipNormals"); } }
 
         public override TooltipContent tooltip
         {
@@ -55,6 +49,7 @@ namespace UnityEditor.ProBuilder.Actions
                 pb.Optimize();
             }
 
+            ProBuilderEditor.Refresh();
             return new ActionResult(ActionResult.Status.Success, "Flip Object Normals");
         }
     }