+++ Editor/MenuActions/Selection/SelectMaterial.cs
@@ -2,7 +2,10 @@ using UnityEngine;
 using System.Collections.Generic;
 using System.Linq;
 using UnityEngine.ProBuilder;
-using UnityEditor.SettingsManagement;
+
+#if UNITY_2023_2_OR_NEWER
+using UnityEngine.UIElements;
+#endif
 
 namespace UnityEditor.ProBuilder.Actions
 {
@@ -16,10 +19,7 @@ namespace UnityEditor.ProBuilder.Actions
             get { return ToolbarGroup.Selection; }
         }
 
-        public override Texture2D icon
-        {
-            get { return IconUtility.GetIcon("Toolbar/Selection_SelectByMaterial", IconSkin.Pro); }
-        }
+        public override Texture2D icon { get { return IconUtility.GetIcon("Toolbar/Selection_SelectByMaterial"); } }
 
         public override TooltipContent tooltip
         {
@@ -46,6 +46,24 @@ namespace UnityEditor.ProBuilder.Actions
         {
             get { return MenuActionState.VisibleAndEnabled; }
         }
+
+#if UNITY_2023_2_OR_NEWER
+        public override VisualElement CreateSettingsContent()
+        {
+            var root = new VisualElement();
+
+            var toggle = new Toggle(gc_restrictToSelection.text);
+            toggle.tooltip = gc_restrictToSelection.tooltip;
+            toggle.SetValueWithoutNotify(m_RestrictToSelectedObjects);
+            toggle.RegisterCallback<ChangeEvent<bool>>(evt =>
+            {
+                m_RestrictToSelectedObjects.SetValue(evt.newValue);
+            });
+            root.Add(toggle);
+
+            return root;
+        }
+#endif
 
         protected override void OnSettingsGUI()
         {