+++ CHANGELOG.md
@@ -15,6 +15,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Significantly improved performance of "ProBuilder Window."
 
+### Changes
+
+- Moved most of ProBuilder actions in the SceneView context menu and overlays when using Unity 2023.2 and newer.
+
+
 ## [5.1.1] - 2023-08-01
 
 ### Fixed
@@ -22,6 +27,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed an exception in `CombineMeshes.AccumulateMeshesInfo` when a `ProBuilderMesh` contains an empty `Material` array.
 - [case: PBLD-78] Fixed a regression where it was no longer possible to set the extrusion point with the Poly Shape tool.
 - [case: UUM-43960] Fixed warning on project import caused by merge conflicts in a meta file.
+
 
 ## [5.1.0] - 2023-06-01